authorization_service/requirements.txt
fastapi==0.70.0
python-jose==3.3.0
uvicorn==0.15.0
